[refactor]Generalize volume handling
The only difference between the volumes used by the NovaAPI and the
NovaConductor is the name of the ConfigMaps the volume mounts. So move
the volume code to a common place and make the ConfigMap names as
parameters to reduce code duplication.
